[REFACTORING] Split IMAPServerPlainAuthDisallowedTest
